+++ inc/mappers/PxMapper.h
@@ -12,5 +12,5 @@ public:
     ~PxMapper();
 
     virtual uint32_t cpuGetMappedAddress(uint16_t address) = 0;
-    virtual uint8_t ppuReadFromCartridge(uint16_t address) = 0;
+    virtual uint32_t ppuReadFromCartridge(uint16_t address) = 0;
 };
